refactor(settings): added more compoentns for ui
