+++ src/components/accounts/ui/index.js
@@ -4,3 +4,4 @@ export * from './menu';
 export * from './table';
 export * from './select';
 export * from './forms';
+export * from './textarea';
