+++ src/components/accounts/ui/textarea/index.js
@@ -1,0 +1,1 @@
+export * from './AccountsTextarea';
